carry the process of src/ast.h

+++ src/lex.h
@@ -113,7 +113,7 @@ typedef struct{
                           // always filled with the capacity of LEX_BUFF_SIZE so
                           // the buff_end_index will flag this situation
 
-    int64      line;       // record the current analyzing line count
+    uint32     line;       // record the current analyzing line count
     lex_token  lextkn;     // to storage the information of the token which is parsing now
     bool       parse_lock; // if the parse_lock == true, the lexical analyzer can not
                            // continue to parse the next token
